requirements.txt

absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.4.0
bitarray==3.7.2
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
ckzg==2.1.5
click==8.3.0
colorama==0.4.6
cytoolz==1.1.0
datasets==4.3.0
dill==0.4.0
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==5.2.0
exceptiongroup==1.3.0
filelock==3.20.0
Flask==3.1.2
flask-cors==6.0.1
flatbuffers==25.9.23
frozenlist==1.8.0
fsspec==2025.9.0
gast==0.4.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.41.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
hexbytes==1.3.1
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.2.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keras==2.11.0
libclang==18.1.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.3
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.16
namex==0.1.0
narwhals==2.9.0
networkx==3.4.2
numpy==1.24.4
oauthlib==3.3.1
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==1.4.4
parsimonious==0.10.0
pillow==11.3.0
platformdirs==4.5.0
propcache==0.4.1
protobuf==3.20.3
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycryptodome==3.23.0
pydantic==2.12.3
pydantic_core==2.41.4
pydeck==0.9.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
pytz==2025.2
pyunormalize==17.0.0
pywin32==311
PyYAML==6.0.3
referencing==0.37.0
regex==2025.10.23
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rlp==4.1.0
rpds-py==0.28.0
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.0.2
scipy==1.15.3
sentencepiece==0.2.1
sentry-sdk==2.42.1
seqeval==1.2.2
simpletransformers==0.70.5
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.50.0
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6.4
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.2.0
threadpoolctl==3.6.0
tokenizers==0.22.1
toml==0.10.2
toolz==1.1.0
torch==2.9.0+cpu
torchaudio==2.9.0+cpu
torchvision==0.24.0+cpu
tornado==6.5.2
tqdm==4.67.1
transformers==4.57.1
types-requests==2.32.4.20250913
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
wandb==0.22.2
watchdog==6.0.0
web3==7.14.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==2.0.0
xxhash==3.6.0
yarl==1.22.0